Refactoring

+++ src/main/java/ru/tink/practice/exception/SecurityNotFoundException.java
@@ -2,7 +2,7 @@ package ru.tink.practice.exception;
 
 public class SecurityNotFoundException extends RuntimeException {
 
-    public SecurityNotFoundException(Integer id) {
+    public SecurityNotFoundException(Long id) {
         super(String.format("Security with %s not found", id));
     }
 }
