+++ src/main/java/ru/tink/practice/repository/PortfolioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.tink.practice.entity.Portfolio;
 
 @Repository
-public interface PortfolioRepository extends JpaRepository<Portfolio, Integer> {
+public interface PortfolioRepository extends JpaRepository<Portfolio, Long> {
 }
